AIILG 443 minor cleanup
